refactor(lib): flatten lib directory structure
Removes lib/infra/ and lib/parsers/ abstraction layers, moving utilities
to lib/ root. Extracts cookie auth from lib/db/user into lib/cookies.ts.
Removes ride mutation functions from lib/db/ride (moved to API routes).
