+++ lib/const.ts
@@ -1,0 +1,1 @@
+export const TIME_ZONE = "America/New_York";
