+++ lib/db/index.ts
@@ -1,13 +1,5 @@
 export { prisma } from "./prisma";
 export {
-  createRide,
-  closeRide,
-  bookmarkRide,
   findManyRides,
-  findBookmarkedRides,
-  findFilteredRides,
+  findBookmarkedRides
 } from "./ride";
-export {
-  getUserFromCookies,
-  getUserNetIdFromCookies,
-} from "./user";
