+++ lib/db/prisma.ts
@@ -9,6 +9,7 @@ const adapter = new PrismaPg({
   connectionString: process.env.DATABASE_URL,
 });
 
+// Set up a singleton class to avoid opening multiple connections on hot reload
 const prismaInstance = globalForPrisma.prisma ?? new PrismaClient({ adapter });
 
 if (process.env.NODE_ENV !== "production") {
