--- lib/infra/index.ts
@@ -1,2 +1,0 @@
-export { withApiErrorHandler, ApiError } from "./apiErrorHandler";
-export { default as logger } from "./logger";
