--- lib/parsers/url.ts
@@ -1,14 +1,0 @@
-import { searchParamsType } from "@/app/interface/main";
-import { decodeDate } from "@/lib/parsers";
-
-// custom type here because nextjs passes params as a plain object
-// results page should be refactored, find a fix for this
-export function decodeSearchParams(searchParams: searchParamsType) {
-  const from = decodeURIComponent(searchParams.from);
-  const to = decodeURIComponent(searchParams.to);
-  const date = decodeDate(searchParams.date);
-  const startTime = decodeURIComponent(searchParams.startTime);
-  const endTime = decodeURIComponent(searchParams.endTime);
-
-  return { from, to, date, startTime, endTime };
-}
